pyproject.toml
[tool.poetry]
name = "events-gateway"
version = "0.1"
description = "Gateway for receiving events from QRadar SIEM system via SYSLOG protocol"
authors = ["Tof <tof.aliev@ru.axept.com>"]
license = "MIT License"
readme = "README.md"
packages = [{include = "events_gateway"}]

[tool.poetry.dependencies]
python = "^3.10"
kafka-python = "^2.0.2"
flake8 = "^5.0.4"
pydantic = "^1.10.2"
loguru = "^0.6.0"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
